+++ packages/components/src/button/Button.js
@@ -3,36 +3,36 @@ import Types from 'prop-types';
 import classNames from 'classnames';
 import './Button.less';
 
-export const ButtonType = {
+const Type = {
   Primary: 'primary',
   Pay: 'pay',
   Secondary: 'secondary',
   Danger: 'danger',
 };
 
-export const ButtonSize = {
+const Size = {
   ExtraSmall: 'xs',
   Small: 'sm',
   Medium: 'md',
   Large: 'lg',
 };
 
-export const ButtonState = {
+const State = {
   Default: 'default',
   Loading: 'loading',
   Disabled: 'disabled',
 };
 
 const Button = ({ label, state, size, type, block, onClick }) => {
-  const isDisabled = state === ButtonState.Disabled;
-  const isLoading = state === ButtonState.Loading;
+  const isDisabled = state === State.Disabled;
+  const isLoading = state === State.Loading;
 
   const classes = classNames(`btn btn-${size}`, {
     'btn-loading': isLoading,
-    'btn-primary': type === ButtonType.Primary,
-    'btn-success': type === ButtonType.Pay,
-    'btn-default': type === ButtonType.Secondary,
-    'btn-danger': type === ButtonType.Danger,
+    'btn-primary': type === Type.Primary,
+    'btn-success': type === Type.Pay,
+    'btn-default': type === Type.Secondary,
+    'btn-danger': type === Type.Danger,
     'btn-block': block,
   });
 
@@ -41,7 +41,7 @@ const Button = ({ label, state, size, type, block, onClick }) => {
       type="button"
       className={classes}
       onClick={e => onClick(e)}
-      disabled={isDisabled || state === ButtonState.Loading}
+      disabled={isDisabled || state === State.Loading}
     >
       {label}
       {isLoading && <span className={classNames('btn-loader', { 'm-l-2': !block })} />}
@@ -49,19 +49,23 @@ const Button = ({ label, state, size, type, block, onClick }) => {
   );
 };
 
+Button.Type = Type;
+Button.Size = Size;
+Button.State = State;
+
 Button.propTypes = {
-  type: Types.oneOf(Object.values(ButtonType)),
-  size: Types.oneOf(Object.values(ButtonSize)),
-  state: Types.oneOf(Object.values(ButtonState)),
+  type: Types.oneOf(Object.values(Type)),
+  size: Types.oneOf(Object.values(Size)),
+  state: Types.oneOf(Object.values(State)),
   block: Types.bool,
   onClick: Types.func.isRequired,
   label: Types.string.isRequired,
 };
 
 Button.defaultProps = {
-  size: ButtonSize.Medium,
-  type: ButtonType.Primary,
-  state: ButtonState.Default,
+  size: Button.Size.Medium,
+  type: Button.Type.Primary,
+  state: Button.State.Default,
   block: false,
 };