+++ packages/components/src/processIndicator/ProcessIndicator.spec.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import { mount } from 'enzyme';
-import { processIndicatorStatuses } from './ProcessIndicator';
 
 import ProcessIndicator from './';
 
@@ -8,8 +7,8 @@ describe('processIndicator', () => {
   let wrapper;
   const ANIMATION_DURATION = 1500;
   const props = {
-    size: 'sm',
-    status: 'processing',
+    status: ProcessIndicator.Status.Processing,
+    size: ProcessIndicator.Size.Small,
     onAnimationCompleted: jest.fn(),
   };
 
@@ -90,12 +89,12 @@ describe('processIndicator', () => {
 
   it('calls onAnimationCompleted with Delay', () => {
     expect(props.onAnimationCompleted).not.toHaveBeenCalled();
-    wrapper.setProps({ status: processIndicatorStatuses[1] });
+    wrapper.setProps({ status: ProcessIndicator.Status.Failed });
     expect(props.onAnimationCompleted).not.toHaveBeenCalled();
     jest.runTimersToTime(ANIMATION_DURATION * 2);
-    expect(props.onAnimationCompleted).toHaveBeenCalledWith(processIndicatorStatuses[1]);
-    wrapper.setProps({ status: processIndicatorStatuses[2] });
+    expect(props.onAnimationCompleted).toHaveBeenCalledWith(ProcessIndicator.Status.Failed);
+    wrapper.setProps({ status: ProcessIndicator.Status.Succeeded });
     jest.runTimersToTime(ANIMATION_DURATION * 2);
-    expect(props.onAnimationCompleted).toHaveBeenCalledWith(processIndicatorStatuses[2]);
+    expect(props.onAnimationCompleted).toHaveBeenCalledWith(ProcessIndicator.Status.Succeeded);
   });
 });